requirements.txt
openai==1.3.5
langchain==0.0.325
python-dotenv==1.0.1
streamlit==1.31.1
faiss-cpu==1.7.4
PyPDF2==3.0.1
tiktoken==0.5.1
